% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{co2emission}
\alias{co2emission}
\title{Amount of CO$_2$ emitted per km

This functions estimates the amount of CO$_2$ emitted (in grams) per kilometer driven
based on the miles per gallon (US gallon) and the fuel type of your car.}
\usage{
co2emission(x, petrol = TRUE)
}
\arguments{
\item{x}{Miles per gallon (measured in US gallons)}

\item{petrol}{Variable to indicate if engine of car uses petrol or diesel fuel}
}
\value{
Grams of CO2 emitted, on average, per kilometer of driving
}
\description{
Amount of CO$_2$ emitted per km

This functions estimates the amount of CO$_2$ emitted (in grams) per kilometer driven
based on the miles per gallon (US gallon) and the fuel type of your car.
}
\examples{
co2emission(25, petrol = TRUE)
co2emissions(42, petrol = FALSE)
}
