\name{emissions-package}

\alias{emissions-package}

\alias{emissions}
\docType{package}

\title{
What the package does (short line)
package title
}

\description{

Provide an overview of how to use the package, including the most important ~~

}

\author{
Joshua Perlman and Robert Hijmans
Maintainer: Who to complain to <yourfault@somewhere.net>
}

