% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Carbon Emissions Calculator}
\usage{
emissions(category, type, value)
}
\arguments{
\item{category}{category to calculate emissions for. Currently supported are:
"airplane" for air travel.}

\item{type}{category-dependent additional information. For air travel: "distance" or "spend".
If you choose "spend", a default conversion factor will be used that maps spend to distance.}

\item{value}{category-dependent value. For air travel, integer vector of either kilometers
traveled or spend in USD. E.g., `value = c(220,800,2500)`.}
}
\value{
CO2 equivalent emissions in tons
}
\description{
Estimate green house gas emissions from travel
}
\examples{
emissions("airplane", "distance", c(100, 1000, 10000))

}
