% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_data.R
\docType{data}
\name{flight_data}
\alias{flight_data}
\title{Sample flight data}
\format{
A data frame with 8 rows and 7 variables.
}
\usage{
flight_data
}
\description{
Sample flight data
}
\keyword{datasets}
