% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{co2emission}
\alias{co2emission}
\title{@title: Amount of Carbon Dioxide emitted per km}
\usage{
co2emission(x, petrol = TRUE)
}
\arguments{
\item{x}{Miles per gallon (measured in US gallons)}

\item{petrol}{Variable to indicate if engine of car uses petrol or diesel fuel}
}
\value{
Grams of CO2 emitted, on average, per kilometer of driving
}
\description{
@title: Amount of Carbon Dioxide emitted per km
}
\examples{
co2emission(25, petrol = TRUE)
co2emission(42, petrol = FALSE)
}
