% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{co2emission}
\alias{co2emission}
\title{: Amount of Carbon Dioxide emitted per km}
\usage{
co2emission(x, petrol = TRUE)
}
\arguments{
\item{x}{An object of class "numeric". This parameter represents miles per gallon (measured in US gallons) of vehicle}

\item{petrol}{A conditional statement to indicate if engine of car uses petrol or diesel fuel.}
}
\value{
An object of class "numeric". Grams of CO2 emitted, on average, per kilometer of driving.
}
\description{
: This functions estimates the amount of Carbon Dioxide emitted (in grams) per kilometer driven
based on the miles per gallon (US gallon) and the fuel type of your car.
}
\examples{
co2emission(25, petrol = TRUE)
co2emission(42, petrol = FALSE)
}
