
\name{emission-package}
\alias{emission-package}
\alias{emission}
\docType{package}

\title{
What the package does (short line)
package title
}

\description{

Provide an overview of how to use the package, including the most important ~~



}

\author{
Who wrote it
Maintainer: Who to complain to <yourfault@somewhere.net>
}

