% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Carbon Emissions Calculator}
\usage{
emissions(category = "airplane", method = "distance", value = 1)
}
\arguments{
\item{category}{category to calculate emissions for. Currently supported are: "airplane" for travel.}

\item{method}{The calculation method to use: 'distance' for distance-based , or 'spend' for spend-based.}

\item{value}{an integer value representing kilometers traveled for distance-based method, or US dollars spent for spend-based method.}
}
\value{
CO2 emissions in tons
}
\description{
Estimate carbon dioxide emissions from travel
}
\examples{
emissions("airplane", "distance", 1000)

}
