% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Carbon Emissions Calculator}
\usage{
emissions(category, value)
}
\arguments{
\item{category}{category to calculate emissions for. Currently supported are:
"airplane" for air travel.}

\item{value}{category-dependent value. For air travel, integer vector of kilometers
traveled. E.g., `value = c(220,800,2500)`.}
}
\value{
CO2 equivalent emissions in tons
}
\description{
Estimate green house gas emissions (CO2e)
}
\examples{
emissions("airplane", c(100, 1000, 10000))

}
