% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Carbon Emissions Calculator}
\usage{
emissions(transport_mode = "plane", kilometers = 1)
}
\arguments{
\item{kilometers}{}
}
\description{
Estimate carbon dioxide emissions from travel
}
\examples{
emissions("plane", 1000)
}
