% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Carbon Emissions Calculator}
\usage{
emissions(transport_mode = "plane", method = "distance", value = 1)
}
\arguments{
\item{transport_mode}{mode of transport: 'plane' for air travel}

\item{method}{The calculation method to use: 'distance' for distance-based , or 'spend' for spend-based.}

\item{value}{an integer value representing kilometers traveled for distance-based method, or US dollars spent for spend-based method.}
}
\description{
Estimate carbon dioxide emissions from travel
}
\examples{
emissions("plane", "distance", 1000)

}
